Here we have a slight cosmetic issue:
For some purposes, it would be useful to ensure that the special_q factor is easily distinguishable in the output file. In earlier versions of the siever, the special_q factor would always come last in the array of found factors. In this version, the optimized TD routines may alter the array by placing repeated small factors at the end. This code is old enough that I don't want to remove the repeated factors, in case someone uses them (msieve does not). So instead, put the special_q factor first and then print the array backward.

git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@412 80ef69ca-9341-0410-9d60-8acc7774fa31