asm "a" constraint doesn't work on gcc 4.5 x86_64; use "m" instead
gnfs-lasieve4I1 was segfaulting on the first movq in optsieve;
gdb showed that rax contained a 1 rather than an address.  The "a"
constraint isn't documented that I can find, and "m" seems to be
what is meant, so switch over.

Successfully factored a 91 digit composite with this change so it
appears to work OK.